+++ cas-server-support-oauth-webflow/build.gradle
@@ -1,0 +1,6 @@
+
+description = "Apereo CAS OAuth Webflow Support"
+dependencies {
+    compile project(":cas-server-core-webflow")
+    compile project(":cas-server-support-oauth")
+}
