+++ cas-server-support-oauth/build.gradle
@@ -3,7 +3,6 @@ dependencies {
     compileOnly libraries.springcloud
     compile project(":cas-server-support-oauth-core")
     compileOnly project(":cas-server-core-tickets")
-    compileOnly project(":cas-server-core-webflow")
     testCompile project(":cas-server-core-util")
     testCompile project(path: ":cas-server-core-authentication", configuration: "tests")
     testCompile project(":cas-server-core-logout")
