+++ cas-server-support-oauth/src/main/java/org/apereo/cas/config/CasOAuthConfiguration.java
@@ -34,14 +34,11 @@ import org.apereo.cas.support.oauth.web.OAuth20CallbackAuthorizeController;
 import org.apereo.cas.support.oauth.web.OAuth20CallbackAuthorizeViewResolver;
 import org.apereo.cas.support.oauth.web.OAuth20ConsentApprovalViewResolver;
 import org.apereo.cas.support.oauth.web.OAuth20ProfileController;
-import org.apereo.cas.support.oauth.web.flow.OAuth20LogoutAction;
-import org.apereo.cas.support.oauth.web.flow.OAuth20LogoutWebflowConfigurer;
 import org.apereo.cas.ticket.ExpirationPolicy;
 import org.apereo.cas.ticket.UniqueTicketIdGenerator;
 import org.apereo.cas.ticket.registry.TicketRegistry;
 import org.apereo.cas.util.DefaultUniqueTicketIdGenerator;
 import org.apereo.cas.validation.ValidationServiceSelectionStrategy;
-import org.apereo.cas.web.flow.CasWebflowConfigurer;
 import org.jasig.cas.client.util.URIBuilder;
 import org.pac4j.cas.client.CasClient;
 import org.pac4j.core.client.RedirectAction;
@@ -51,7 +48,6 @@ import org.pac4j.core.credentials.authenticator.UsernamePasswordAuthenticator;
 import org.pac4j.core.http.CallbackUrlResolver;
 import org.pac4j.http.client.direct.DirectBasicAuthClient;
 import org.pac4j.http.client.direct.DirectFormClient;
-import org.pac4j.springframework.web.ApplicationLogoutController;
 import org.pac4j.springframework.web.CallbackController;
 import org.pac4j.springframework.web.SecurityInterceptor;
 import org.springframework.beans.factory.annotation.Autowire;
@@ -65,8 +61,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
-import org.springframework.webflow.definition.registry.FlowDefinitionRegistry;
-import org.springframework.webflow.engine.builder.support.FlowBuilderServices;
 
 import javax.annotation.PostConstruct;
 import javax.servlet.http.HttpServletRequest;
@@ -111,17 +105,7 @@ public class CasOAuthConfiguration extends WebMvcConfigurerAdapter {
     @Qualifier("ticketRegistry")
     private TicketRegistry ticketRegistry;
 
-    @Autowired
-    @Qualifier("loginFlowRegistry")
-    private FlowDefinitionRegistry loginFlowDefinitionRegistry;
 
-    @Autowired
-    @Qualifier("logoutFlowRegistry")
-    private FlowDefinitionRegistry logoutFlowDefinitionRegistry;
-    
-    @Autowired
-    private FlowBuilderServices flowBuilderServices;
-    
     @ConditionalOnMissingBean(name = "accessTokenResponseGenerator")
     @Bean(autowire = Autowire.BY_NAME)
     public AccessTokenResponseGenerator accessTokenResponseGenerator() {
@@ -185,13 +169,13 @@ public class CasOAuthConfiguration extends WebMvcConfigurerAdapter {
             return url;
         };
     }
-    
+
     @ConditionalOnMissingBean(name = "requiresAuthenticationAuthorizeInterceptor")
     @Bean
     public SecurityInterceptor requiresAuthenticationAuthorizeInterceptor() {
         return new SecurityInterceptor(oauthSecConfig(), CAS_OAUTH_CLIENT);
     }
-    
+
     @ConditionalOnMissingBean(name = "consentApprovalViewResolver")
     @Bean
     public ConsentApprovalViewResolver consentApprovalViewResolver() {
@@ -204,30 +188,12 @@ public class CasOAuthConfiguration extends WebMvcConfigurerAdapter {
         return new OAuth20CallbackAuthorizeViewResolver() {
         };
     }
-
-    @Bean
-    public CasWebflowConfigurer oauth20LogoutWebflowConfigurer() {
-        final OAuth20LogoutWebflowConfigurer c = new OAuth20LogoutWebflowConfigurer();
-        c.setFlowBuilderServices(this.flowBuilderServices);
-        c.setLoginFlowDefinitionRegistry(this.loginFlowDefinitionRegistry);
-        c.setLogoutFlowDefinitionRegistry(this.logoutFlowDefinitionRegistry);
-        return c;
-    }
-
-    @Bean
-    public OAuth20LogoutAction oauth20LogoutAction() {
-        final ApplicationLogoutController controller = new ApplicationLogoutController();
-        controller.setConfig(oauthSecConfig());
-        final OAuth20LogoutAction action = new OAuth20LogoutAction();
-        action.setApplicationLogoutController(controller);
-        return action;
-    }
     
     @Bean
     public SecurityInterceptor requiresAuthenticationAccessTokenInterceptor() {
         return new SecurityInterceptor(oauthSecConfig(), "clientBasicAuth,clientForm,userForm");
     }
-    
+
     @Bean
     public HandlerInterceptorAdapter oauthInterceptor() {
         return new HandlerInterceptorAdapter() {
@@ -411,8 +377,8 @@ public class CasOAuthConfiguration extends WebMvcConfigurerAdapter {
         c.setConfig(oauthSecConfig());
         return c;
     }
-    
-    
+
+
     @Bean
     public UniqueTicketIdGenerator accessTokenIdGenerator() {
         return new DefaultUniqueTicketIdGenerator();
