+++ settings.gradle
@@ -67,6 +67,7 @@ include "cas-server-support-x509"
 include "cas-server-support-x509-webflow"
 include "cas-server-support-oauth-core"
 include "cas-server-support-oauth"
+include "cas-server-support-oauth-webflow"
 include "cas-server-support-oidc-core"
 include "cas-server-support-oidc"
 include "cas-server-support-pac4j"
